Switch CSS-parser to crass
The other parser had some issues that I was unable to solve.

This commit fixes most of the CSS-issues, like:
* pseudo-selectors
* pseudo-elements
* has()-selectors
* ... and all the other issues there were with the previous one...

composes works, but is not perfect.. there should be some validations..

The formatter leaves a lot of whitespace.
It probably doesn't matter though because the file sizes are so small.

The source map doesn't have any locations, just the source.
I don't even know if the previous implementation worked properly.