Move %body to root.haml
There is an issue where, if the component that renders the body
element gets updated, then the body element will be removed and
then added again, but there are some issues somewhere, the DOM
gets broken anyways and the page has to be reloaded.

This commit attempts to fix this by moving %body to root.haml,
but it still doesn't work properly. It seems like the previous
layout is still there in the DOM and I don't have the energy to
debug it more right now..