css: Store assets in a set to get the correct order
This makes root.css load first, then layouts etc,
so that styles can be overridden...

Layer order will still be broken in development mode, because new stylesheets
will always be applied at the end. Not sure how to fix this.
Maybe it is possible to not use @layer and instead just keep
the <link> elements ordered? I don't know if the order matters
or of it's the load order that decides the winner.

Related to #38